Add Dial Button Action to StreamDeck vJoy plugin
The commit includes a new Dial Button Action that allows the StreamDeck dial to push vJoy buttons when rotated. To facilitate this feature, a new file DialButtonAction.cs has been added, alongside updates to the project file and manifest. The user interface updates are reflected in the addition of DialButtonAction.html to the PropertyInspector.