Update icons and add install.bat
This commit replaces the existing icons with a new joystick icon and adds the install.bat script with corresponding changes in the csproj, manifest.json, Install (DEBUG).run.xml files. A reference to the icon's source has been added to the README and the new icon file, joystick-svgrepo-com.svg, has been added to the repository.